+++ lib/src/microsoft/voices/voice_model.dart
@@ -1,11 +1,10 @@
 import 'dart:ui';
 
+import 'package:cloud_text_to_speech/src/common/locale/locale_helpers.dart';
 import 'package:cloud_text_to_speech/src/common/tts/tts_providers.dart';
-import 'package:cloud_text_to_speech/src/common/utils/helpers.dart';
 import 'package:cloud_text_to_speech/src/common/locale/locale_model.dart';
 import 'package:cloud_text_to_speech/src/universal/voices/voice_model.dart';
 import 'package:json_annotation/json_annotation.dart';
-import 'package:locale_names/locale_names.dart';
 
 part 'voice_model.g.dart';
 
@@ -70,20 +69,8 @@ class VoiceMicrosoft extends VoiceUniversal {
   static VoiceLocale _toLocale(String locale) {
     List<String> localeSegments = locale.split('-');
 
-    Locale localeObj = Helpers.segmentsToLocale(localeSegments);
+    Locale localeObj = VoiceLocaleHelpers.segmentsToLocale(localeSegments);
 
-    return VoiceLocale(
-        code: localeSegments.join('-'),
-        name: Helpers.formatLanguageCountry(
-            localeObj.defaultDisplayLanguage, localeObj.defaultDisplayCountry),
-        nativeName: Helpers.formatLanguageCountry(
-            localeObj.nativeDisplayLanguage, localeObj.nativeDisplayCountry),
-        languageCode: localeObj.languageCode,
-        languageName: localeObj.defaultDisplayLanguage,
-        nativeLanguageName: localeObj.nativeDisplayLanguage,
-        countryCode: localeObj.countryCode,
-        countryName: localeObj.defaultDisplayCountry,
-        nativeCountryName: localeObj.nativeDisplayCountry,
-        scriptCode: localeObj.scriptCode);
+    return VoiceLocaleHelpers.localeToVoiceLocale(localeObj);
   }
 }